Revert "Genericized the StationUpcomingTableViewCell into StationUpcomingCell protocol."
This reverts commit 72b65227947669110d14ac17a4c3eb858b6731f4.
