--- Proper/UI/StationUpcomingCell.xib
@@ -1,96 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11762" systemVersion="16B2555" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11757"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="StationUpcomingTableViewCell" customModule="Proper" customModuleProvider="target">
-            <connections>
-                <outlet property="collectionView" destination="8v6-7d-hPA" id="yG9-eS-HhR"/>
-                <outlet property="subtitle" destination="mdB-Z2-4Lm" id="vjQ-N9-4Bw"/>
-                <outlet property="title" destination="pgQ-9z-u6k" id="l9H-Iu-Kjy"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" accessoryType="disclosureIndicator" indentationWidth="10" rowHeight="90" id="g39-MD-gHE" customClass="StationUpcomingTableViewCell" customModule="Proper" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="400" height="90"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="g39-MD-gHE" id="Ktx-4F-rJz">
-                <rect key="frame" x="0.0" y="0.0" width="367" height="89"/>
-                <autoresizingMask key="autoresizingMask"/>
-            </tableViewCellContentView>
-            <point key="canvasLocation" x="349" y="364"/>
-        </tableViewCell>
-        <view contentMode="scaleToFill" id="mi2-fW-CcY">
-            <rect key="frame" x="0.0" y="0.0" width="367" height="89"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <subviews>
-                <label opaque="NO" multipleTouchEnabled="YES" contentMode="left" text="Title" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="pgQ-9z-u6k" userLabel="Title" customClass="TransitLabel" customModule="Proper" customModuleProvider="target">
-                    <rect key="frame" x="7" y="9" width="32" height="20"/>
-                    <constraints>
-                        <constraint firstAttribute="width" constant="442" id="tRW-PV-1JB"/>
-                        <constraint firstAttribute="width" relation="greaterThanOrEqual" constant="90" id="uUT-E1-pdT"/>
-                    </constraints>
-                    <fontDescription key="fontDescription" type="system" pointSize="16"/>
-                    <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                    <nil key="highlightedColor"/>
-                    <variation key="default">
-                        <mask key="constraints">
-                            <exclude reference="uUT-E1-pdT"/>
-                            <exclude reference="tRW-PV-1JB"/>
-                        </mask>
-                    </variation>
-                </label>
-                <label opaque="NO" multipleTouchEnabled="YES" contentMode="left" text="Subtitle" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="mdB-Z2-4Lm" customClass="TransitLabel" customModule="Proper" customModuleProvider="target">
-                    <rect key="frame" x="7" y="29" width="41" height="14"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="11"/>
-                    <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" scrollEnabled="NO" showsHorizontalScrollIndicator="NO" showsVerticalScrollIndicator="NO" delaysContentTouches="NO" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="8v6-7d-hPA">
-                    <rect key="frame" x="7" y="50" width="353" height="30"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="30" id="Irk-Sv-LOm"/>
-                    </constraints>
-                    <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" minimumLineSpacing="10" minimumInteritemSpacing="10" id="hAD-Cv-Tji">
-                        <size key="itemSize" width="30" height="30"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                </collectionView>
-            </subviews>
-            <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-            <constraints>
-                <constraint firstItem="8v6-7d-hPA" firstAttribute="top" secondItem="mdB-Z2-4Lm" secondAttribute="bottom" constant="7" id="JBU-VS-EKZ"/>
-                <constraint firstItem="mdB-Z2-4Lm" firstAttribute="top" secondItem="pgQ-9z-u6k" secondAttribute="bottom" id="VBe-EB-XaQ"/>
-                <constraint firstItem="pgQ-9z-u6k" firstAttribute="leading" secondItem="mi2-fW-CcY" secondAttribute="leading" constant="7" id="beH-fW-Z9A"/>
-                <constraint firstItem="pgQ-9z-u6k" firstAttribute="top" secondItem="mi2-fW-CcY" secondAttribute="top" constant="9" id="cqO-ii-CsD"/>
-                <constraint firstItem="8v6-7d-hPA" firstAttribute="leading" secondItem="mi2-fW-CcY" secondAttribute="leading" constant="7" id="htX-qd-u9P"/>
-                <constraint firstAttribute="trailing" secondItem="8v6-7d-hPA" secondAttribute="trailing" constant="7" id="iTt-nf-dHg"/>
-                <constraint firstItem="mdB-Z2-4Lm" firstAttribute="leading" secondItem="mi2-fW-CcY" secondAttribute="leading" constant="7" id="rwy-Gw-09w"/>
-            </constraints>
-            <nil key="simulatedStatusBarMetrics"/>
-            <nil key="simulatedTopBarMetrics"/>
-            <nil key="simulatedBottomBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <point key="canvasLocation" x="333" y="177"/>
-        </view>
-        <view contentMode="scaleToFill" id="agU-je-AiO" customClass="UITableViewHeaderFooterView">
-            <rect key="frame" x="0.0" y="0.0" width="400" height="90"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-            <nil key="simulatedStatusBarMetrics"/>
-            <nil key="simulatedTopBarMetrics"/>
-            <nil key="simulatedBottomBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <point key="canvasLocation" x="349" y="531"/>
-        </view>
-    </objects>
-</document>
