+++ Proper/UI/StationUpcomingTableViewCell.xib
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11762" systemVersion="16B2555" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11757"/>
+        <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell contentMode="scaleToFill" selectionStyle="default" accessoryType="disclosureIndicator" indentationWidth="10" rowHeight="90" id="g39-MD-gHE" customClass="StationUpcomingTableViewCell" customModule="Proper" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="400" height="90"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="g39-MD-gHE" id="Ktx-4F-rJz">
+                <rect key="frame" x="0.0" y="0.0" width="367" height="89"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <label opaque="NO" multipleTouchEnabled="YES" contentMode="left" text="Title" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="HNo-Xa-pEl" userLabel="Title" customClass="TransitLabel" customModule="Proper" customModuleProvider="target">
+                        <rect key="frame" x="15" y="5" width="32" height="20"/>
+                        <constraints>
+                            <constraint firstAttribute="width" relation="greaterThanOrEqual" constant="90" id="378-Mg-EB9"/>
+                            <constraint firstAttribute="width" constant="442" id="cjf-oz-tQp"/>
+                        </constraints>
+                        <fontDescription key="fontDescription" type="system" pointSize="16"/>
+                        <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                        <nil key="highlightedColor"/>
+                        <variation key="default">
+                            <mask key="constraints">
+                                <exclude reference="cjf-oz-tQp"/>
+                                <exclude reference="378-Mg-EB9"/>
+                            </mask>
+                        </variation>
+                    </label>
+                    <label opaque="NO" multipleTouchEnabled="YES" contentMode="left" text="Subtitle" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="gMX-gL-jnt" customClass="TransitLabel" customModule="Proper" customModuleProvider="target">
+                        <rect key="frame" x="15" y="25" width="41" height="14"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="11"/>
+                        <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" fixedFrame="YES" scrollEnabled="NO" showsHorizontalScrollIndicator="NO" showsVerticalScrollIndicator="NO" delaysContentTouches="NO" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="yl3-WI-tGU">
+                        <rect key="frame" x="15" y="47" width="352" height="34"/>
+                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" minimumLineSpacing="10" minimumInteritemSpacing="10" id="tLD-ha-DHg">
+                            <size key="itemSize" width="30" height="30"/>
+                            <size key="headerReferenceSize" width="0.0" height="0.0"/>
+                            <size key="footerReferenceSize" width="0.0" height="0.0"/>
+                            <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
+                        </collectionViewFlowLayout>
+                    </collectionView>
+                </subviews>
+                <constraints>
+                    <constraint firstItem="HNo-Xa-pEl" firstAttribute="leading" secondItem="Ktx-4F-rJz" secondAttribute="leading" constant="15" id="5M8-5X-Ojj"/>
+                    <constraint firstItem="HNo-Xa-pEl" firstAttribute="top" secondItem="Ktx-4F-rJz" secondAttribute="topMargin" constant="-3" placeholder="YES" id="F07-cr-o0v"/>
+                    <constraint firstItem="gMX-gL-jnt" firstAttribute="top" secondItem="HNo-Xa-pEl" secondAttribute="bottom" placeholder="YES" id="XI0-RZ-kY4"/>
+                    <constraint firstItem="gMX-gL-jnt" firstAttribute="top" secondItem="HNo-Xa-pEl" secondAttribute="bottom" id="cMS-Ob-i9R"/>
+                    <constraint firstItem="HNo-Xa-pEl" firstAttribute="top" secondItem="Ktx-4F-rJz" secondAttribute="topMargin" constant="-3" id="vLg-40-10Y"/>
+                    <constraint firstItem="gMX-gL-jnt" firstAttribute="leading" secondItem="Ktx-4F-rJz" secondAttribute="leading" constant="15" id="ytm-04-tMG"/>
+                </constraints>
+                <variation key="default">
+                    <mask key="constraints">
+                        <exclude reference="F07-cr-o0v"/>
+                        <exclude reference="XI0-RZ-kY4"/>
+                    </mask>
+                </variation>
+            </tableViewCellContentView>
+            <connections>
+                <outlet property="collectionView" destination="yl3-WI-tGU" id="qz2-Xe-R7c"/>
+                <outlet property="subtitle" destination="gMX-gL-jnt" id="9L6-Wp-yBC"/>
+                <outlet property="title" destination="HNo-Xa-pEl" id="YcC-ri-SLl"/>
+            </connections>
+            <point key="canvasLocation" x="349" y="364"/>
+        </tableViewCell>
+    </objects>
+</document>
